1st set of unit tests working. Updated CAProtocol to give allOn optional duration argument
