+++ STM32/DC/Core/Src/DCBoard.c
@@ -27,7 +27,7 @@
 #define MAX_PWM TURNONPWM
 
 // Private static function declarations.
-static void CAallOn(bool isOn);
+static void CAallOn(bool isOn, int duration_ms);
 static void CAportState(int port, bool state, int percent, int duration);
 
 // Local variables.
@@ -222,8 +222,11 @@ static void actuatePins(ActuationInfo actuationInfo)
     }
 }
 
-static void CAallOn(bool isOn)
+static void CAallOn(bool isOn, int duration_ms)
 {
+    /* Unused argument required for compatibility with AC board */
+    (void) duration_ms;
+
     (isOn) ? allOn() : allOff();
 }
 static void CAportState(int port, bool state, int percent, int duration)
